fix/Make BlogList component reusable
close #137
