+++ client/src/components/BlogList/index.jsx
@@ -1,29 +1,11 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React from 'react';
+import PropTypes from 'prop-types';
 
-import axiosConfig from '../../config/axiosConfig';
 import BlogCard from '../BlogCard';
 import LoaderCentered from '../../elements/LoaderCentered/LoaderCenter';
 import { defaultContent, defaultImage } from '../../config/constants';
 
-export default function BlogList() {
-  const [blogs, setBlogs] = useState([]);
-
-  const fetchBlogs = async () => {
-    const params = {
-      limit: 10,
-      skip: 0,
-      orderBy: 'createdAt',
-      orderType: -1,
-    };
-    const response = await axios.get(`posts?limit=${params.limit}&skip=${params.skip}&orderby=${params.orderBy}&orderType=${params.orderType}`, axiosConfig);
-    setBlogs(response.data.posts);
-  };
-
-  useEffect(() => {
-    fetchBlogs();
-  }, []);
-
+export default function BlogList({ blogs }) {
   const conditionallyRenderBlogs = () => {
     if (blogs.length === 0) {
       return <LoaderCentered />;
@@ -44,3 +26,23 @@ export default function BlogList() {
     </div>
   );
 }
+
+BlogList.defaultProps = {
+  blogs: [],
+};
+
+BlogList.propTypes = {
+  blogs: PropTypes.arrayOf(
+    PropTypes.shape({
+      author: PropTypes.string,
+      category: PropTypes.string,
+      content: PropTypes.string,
+      public: PropTypes.bool,
+      published: PropTypes.bool,
+      title: PropTypes.string,
+      upvotes: PropTypes.array,
+      views: PropTypes.array,
+      _id: PropTypes.string,
+    }),
+  ),
+};
